update 4.x-develop with changes from 3.x-develop

+++ tests/goss/run_all.sh
@@ -10,6 +10,11 @@ for file in $TEST_DIR/goss.*.yaml; do
   prefix=$TEST_DIR/goss.
   service=${file/$prefix/}
   service=${service/.yaml/}
+  # @todo: Re-enable solr tests once goss bytes.Reader bug is resolved.
+  if [ "$service" == "solr" ]; then
+    echo "==> Skipping tests for \"$service\" service (goss bug)"
+    continue
+  fi
   echo "==> Running tests for \"$service\" service"
   goss -g $file render > goss.yaml && dgoss run -i -e NGINX_FASTCGI_PASS=localhost $DOCKERHUB_NAMESPACE/$service:$GOVCMS_RELEASE_TAG || ((fails++))
   rm -Rf goss.yaml
